review rework 2: unexport agentgateway's Engine and merge into one file
Only gateway.Engine is a public contract; the concrete implementation
struct is now unexported (engine), with NewEngine still returning the
interface. Also folded engine.go into agentgateway.go since both files
held pieces of the same single implementation type.
